Transfrom tree integration & simple stupid frustum culling

+++ Engine/DirectX11_Starter/Object.cpp
@@ -12,7 +12,7 @@ Object::Object()
 
 Object::~Object()
 {
-	for (int i = 0; i < allObjects.size(); i++) {
+	for (size_t i = 0; i < allObjects.size(); i++) {
 		if (allObjects[i].first == id) {
 			allObjects.erase(allObjects.begin() + i);
 			break;
